Made Zinzinnati a Merchant Republic among other things.
Zinzinnati is now a Merchant Republic.
The Ohio River is now fully navigable by everyone.
Changed the Aurora dynasty to Pulaski.
Changed the Ashland dynasty to Sheridan.
Changed the Ashford dynasty to Ashland.
Added an Archer dynasty to the history of Joliet.
CWC now has an event that makes him OP, but it only fires if the Absurd Events game rule is active.